// src/content/searchItems.interface.ts
export type SearchItem = {
  _source: {
    language: number;
    manufacturer: string;
    model: string;
    name: string;
    name_language_3: string;
    phonefk: number;
    pk: number;
    type: string;
    val?: string;
    val_language_3?: string;
    steps_language_1?: string;
    steps_language_3?: string;
    concept: string;
    phone_type: string;
  };
}